fix(uploader): Add slash to minidump upload URL (#269)
The initial version of the Electron SDK used to send minidumps to an
endpoint without the trailing slash since the server initially did not
support this by mistake. This has long been fixed, and the current
version of the SDK even requires the Envelope endpoint for most of its
operation.